requirements.txt
# Stock API, Backtest
numpy==2.2.6
openpyxl==3.1.2
pandas==2.2.0
PyYAML==6.0
requests==2.31.0
tqdm==4.66.0
XlsxWriter==3.2.0
python-dotenv==1.2.2

# RL
torch==2.2
gymnasium==1.2.3
stable-baselines3==2.8.0
pandas-ta==0.4.71b0
